Do not allow line-breaks for non-word character groupings.
This adds a new step of a "render" of the data into view, which allows us to modify the representation via innerHTML. We use this step to perform the following algorithm:

* Split text into "chunks", delineated by whitespace
* Convert all short "chunks" into `<nobr>` elements.
* Reassemble all chunks.

This addition presents challenges to the cursor state maintenance. Previously it assumed that `#text` always had exactly one text element, but now it might contain `<nobr>` elements. In order to handle this, a new function is added that converts cursor state back to something the browser expects post-render.

Fixes #1